optional MathJax
Mathjax is awesome but I needed to address an issue with page loading
(also SEO) on pages on the site that didn’t utilize MathJax, but still
loaded the script. This was bad.

I changed how adding MathJax to pages works. To specify mathjax in a
post then a property must be added to the YAML front matter in a post
something simply like

```
—
mathjax: true
—
```

will allow the mathjax script to be added when the site is built with
jekyll.